Descriptions and change category_group name
The descriptions of the variables have been taken out and put into a separate file which I am going to upload onto this site and update in the shared google doc. They should be in their own separate document. I changed 'transportation' to 'parking_transportation' because parking is included in the grouping and I thought that was important to include in the name. You can change it back if you want.